correcciones templeates docentes y alumnos + cambio tipo de dato smallint a int

+++ database/migrations/2022_10_01_191905_create_carreras_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('carreras', function (Blueprint $table) {
             $table->id();
-            $table->unsignedSmallInteger('codigo');
+            $table->unsignedInteger('codigo');
             $table->text('nombre');
             
             $table->unsignedBigInteger('id_departamento');
