+++ database/migrations/2022_10_01_193629_create_alumnos_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
         Schema::create('alumnos', function (Blueprint $table) {
             $table->id();
             $table->string('nombre');
-            $table->unsignedSmallInteger('lu')->unique();
-            $table->unsignedSmallInteger('dni')->unique();
+            $table->unsignedInteger('lu')->unique();
+            $table->unsignedInteger('dni')->unique();
             $table->date('fecha de nacimiento');
             $table->string('mail')->unique();
