+++ database/migrations/2022_10_01_194940_create_cursados_notas_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
     {
         Schema::create('cursados_notas', function (Blueprint $table) {
             $table->id();
-            $table->unsignedSmallInteger('nota');
-            $table->unsignedSmallInteger('estado');
+            $table->unsignedInteger('nota');
+            $table->unsignedInteger('estado');
             $table->unsignedBigInteger('id_comision');
             $table->timestamps();
